Add shared-handle, in-buffer, and out-buffer types
This commit adds three new types to the type grammar of interface types.
The first is a `shared-handle` type which is intended to be used for an
unforgeable reference to an externally-defined resource. The
`shared-handle` type is similar to `externref` but there can be multiple
types of `shared-handle` types that don't all typecheck to the same
type. This allows modules to define their own unique resources that
can't be intermingled with other resources.

The other two types are `in-buffer` and `out-buffer`. These types are
intended to roughly represent views into the caller's memory when passed
to the callee. These views are temporary and only last for the duration
of the function call they were passed as paramters to. They also have an
optional capacity for the callee to learn about how many items can
either be read or written.

The motivation for the addition of these types is to move the WASI
specification to using purely interface types. By using pure interface
types it enables WASI to have more intelligent code generators which
understand the purpose of all arguments as well as ensuring that all
arguments and return values are encapsulated enough to work well when
working with multiple modules. For example with these types it would
then be possible to virtualize WASI trivially whereas previously it
required a number of hoops to jump through to make sure pointers are
always relative to the right memory and such.

The text describing these types is not as exhaustive as the surrounding
text for other adapter instructions, but this is somewhat intentional.
The goal of this PR is to get broad consensus on the semantics of these
types and how they're intended to be used. The actual nitty-gritty of
how they can be instantiated and such is thought to be possible to leave
to a future PR/design.